+++ jax_lensing_profiles/MassModel/Profiles/Sersic_ellipse_kappa_jax.py
