+++ jax_lensing_profiles/MassModel/Profiles/__init__.py
@@ -5,4 +5,4 @@ from .multi_gaussian_ellipse_kappa_jax import MultiGaussianEllipseKappa
 from .NFW import NFW
 from .NFW import TNFW
 from .NFW_ellipse_kappa_jax import NFWEllipseKappa
-from .Sersic import SersicEllipseKappa
+from .Sersic_ellipse_kappa_jax import SersicEllipseKappa
